+++ src/netflow_common.rs
@@ -5,7 +5,7 @@ use crate::NetflowPacket;
 use crate::protocol::ProtocolTypes;
 use crate::static_versions::{v5::V5, v7::V7};
 use crate::variable_versions::data_number::FieldValue;
-use crate::variable_versions::ipfix_lookup::IPFixField;
+use crate::variable_versions::ipfix_lookup::{IANAIPFixField, IPFixField};
 use crate::variable_versions::v9_lookup::V9Field;
 use crate::variable_versions::{
     ipfix::{FlowSetBody as IPFixFlowSetBody, IPFix},
@@ -125,7 +125,7 @@ impl From<&V9> for NetflowCommon {
             if let V9FlowSetBody::Data(data) = &flowset.body {
                 for data_field in &data.fields {
                     let value_map: BTreeMap<V9Field, FieldValue> =
-                        data_field.values().cloned().collect();
+                        data_field.clone().into_iter().collect();
                     flowsets.push(NetflowCommonFlowSet {
                         src_addr: value_map
                             .get(&V9Field::Ipv4SrcAddr)
@@ -184,43 +184,50 @@ impl From<&IPFix> for NetflowCommon {
             if let IPFixFlowSetBody::Data(data) = &flowset.body {
                 for data_field in &data.fields {
                     let value_map: BTreeMap<IPFixField, FieldValue> =
-                        data_field.values().cloned().collect();
+                        data_field.clone().into_iter().collect();
                     flowsets.push(NetflowCommonFlowSet {
                         src_addr: value_map
-                            .get(&IPFixField::SourceIpv4address)
-                            .or_else(|| value_map.get(&IPFixField::SourceIpv6address))
+                            .get(&IPFixField::IANA(IANAIPFixField::SourceIpv4address))
+                            .or_else(|| {
+                                value_map
+                                    .get(&IPFixField::IANA(IANAIPFixField::SourceIpv6address))
+                            })
                             .and_then(|v| v.try_into().ok()),
                         dst_addr: value_map
-                            .get(&IPFixField::DestinationIpv4address)
-                            .or_else(|| value_map.get(&IPFixField::DestinationIpv6address))
+                            .get(&IPFixField::IANA(IANAIPFixField::DestinationIpv4address))
+                            .or_else(|| {
+                                value_map.get(&IPFixField::IANA(
+                                    IANAIPFixField::DestinationIpv6address,
+                                ))
+                            })
                             .and_then(|v| v.try_into().ok()),
                         src_port: value_map
-                            .get(&IPFixField::SourceTransportPort)
+                            .get(&IPFixField::IANA(IANAIPFixField::SourceTransportPort))
                             .and_then(|v| v.try_into().ok()),
                         dst_port: value_map
-                            .get(&IPFixField::DestinationTransportPort)
+                            .get(&IPFixField::IANA(IANAIPFixField::DestinationTransportPort))
                             .and_then(|v| v.try_into().ok()),
                         protocol_number: value_map
-                            .get(&IPFixField::ProtocolIdentifier)
+                            .get(&IPFixField::IANA(IANAIPFixField::ProtocolIdentifier))
                             .and_then(|v| v.try_into().ok()),
-                        protocol_type: value_map.get(&IPFixField::ProtocolIdentifier).and_then(
-                            |v| {
+                        protocol_type: value_map
+                            .get(&IPFixField::IANA(IANAIPFixField::ProtocolIdentifier))
+                            .and_then(|v| {
                                 v.try_into()
                                     .ok()
                                     .map(|proto: u8| ProtocolTypes::from(proto))
-                            },
-                        ),
+                            }),
                         first_seen: value_map
-                            .get(&IPFixField::FlowStartSysUpTime)
+                            .get(&IPFixField::IANA(IANAIPFixField::FlowStartSysUpTime))
                             .and_then(|v| v.try_into().ok()),
                         last_seen: value_map
-                            .get(&IPFixField::FlowEndSysUpTime)
+                            .get(&IPFixField::IANA(IANAIPFixField::FlowEndSysUpTime))
                             .and_then(|v| v.try_into().ok()),
                         src_mac: value_map
-                            .get(&IPFixField::SourceMacaddress)
+                            .get(&IPFixField::IANA(IANAIPFixField::SourceMacaddress))
                             .and_then(|v| v.try_into().ok()),
                         dst_mac: value_map
-                            .get(&IPFixField::DestinationMacaddress)
+                            .get(&IPFixField::IANA(IANAIPFixField::DestinationMacaddress))
                             .and_then(|v| v.try_into().ok()),
                     });
                 }
@@ -237,8 +244,6 @@ impl From<&IPFix> for NetflowCommon {
 
 #[cfg(test)]
 mod common_tests {
-
-    use std::collections::BTreeMap;
     use std::net::{IpAddr, Ipv4Addr};
 
     use crate::netflow_common::NetflowCommon;
@@ -249,7 +254,7 @@ mod common_tests {
         Data as IPFixData, FlowSet as IPFixFlowSet, FlowSetBody as IPFixFlowSetBody,
         FlowSetHeader as IPFixFlowSetHeader, Header as IPFixHeader, IPFix,
     };
-    use crate::variable_versions::ipfix_lookup::IPFixField;
+    use crate::variable_versions::ipfix_lookup::{IANAIPFixField, IPFixField};
     use crate::variable_versions::v9::{
         Data as V9Data, FlowSet as V9FlowSet, FlowSetBody as V9FlowSetBody,
         FlowSetHeader as V9FlowSetHeader, Header as V9Header, V9,
@@ -402,66 +407,39 @@ mod common_tests {
                 },
                 body: V9FlowSetBody::Data(V9Data {
                     padding: vec![],
-                    fields: vec![BTreeMap::from([
+                    fields: vec![Vec::from([
                         (
-                            0,
-                            (
-                                V9Field::Ipv4SrcAddr,
-                                FieldValue::Ip4Addr(Ipv4Addr::new(192, 168, 1, 1)),
-                            ),
+                            V9Field::Ipv4SrcAddr,
+                            FieldValue::Ip4Addr(Ipv4Addr::new(192, 168, 1, 1)),
                         ),
                         (
-                            1,
-                            (
-                                V9Field::Ipv4DstAddr,
-                                FieldValue::Ip4Addr(Ipv4Addr::new(192, 168, 1, 2)),
-                            ),
+                            V9Field::Ipv4DstAddr,
+                            FieldValue::Ip4Addr(Ipv4Addr::new(192, 168, 1, 2)),
                         ),
                         (
-                            2,
-                            (
-                                V9Field::L4SrcPort,
-                                FieldValue::DataNumber(DataNumber::U16(1234)),
-                            ),
+                            V9Field::L4SrcPort,
+                            FieldValue::DataNumber(DataNumber::U16(1234)),
                         ),
                         (
-                            3,
-                            (
-                                V9Field::L4DstPort,
-                                FieldValue::DataNumber(DataNumber::U16(80)),
-                            ),
+                            V9Field::L4DstPort,
+                            FieldValue::DataNumber(DataNumber::U16(80)),
+                        ),
+                        (V9Field::Protocol, FieldValue::DataNumber(DataNumber::U8(6))),
+                        (
+                            V9Field::FirstSwitched,
+                            FieldValue::DataNumber(DataNumber::U32(100)),
                         ),
                         (
-                            4,
-                            (V9Field::Protocol, FieldValue::DataNumber(DataNumber::U8(6))),
+                            V9Field::LastSwitched,
+                            FieldValue::DataNumber(DataNumber::U32(200)),
                         ),
                         (
-                            5,
-                            (
-                                V9Field::FirstSwitched,
-                                FieldValue::DataNumber(DataNumber::U32(100)),
-                            ),
+                            V9Field::InSrcMac,
+                            FieldValue::MacAddr("00:00:00:00:00:01".to_string()),
                         ),
                         (
-                            6,
-                            (
-                                V9Field::LastSwitched,
-                                FieldValue::DataNumber(DataNumber::U32(200)),
-                            ),
-                        ),
-                        (
-                            7,
-                            (
-                                V9Field::InSrcMac,
-                                FieldValue::MacAddr("00:00:00:00:00:01".to_string()),
-                            ),
-                        ),
-                        (
-                            8,
-                            (
-                                V9Field::InDstMac,
-                                FieldValue::MacAddr("00:00:00:00:00:02".to_string()),
-                            ),
+                            V9Field::InDstMac,
+                            FieldValue::MacAddr("00:00:00:00:00:02".to_string()),
                         ),
                     ])],
                 }),
@@ -511,70 +489,42 @@ mod common_tests {
                     length: 0,
                 },
                 body: IPFixFlowSetBody::Data(IPFixData {
-                    padding: vec![],
-                    fields: vec![BTreeMap::from([
+                    fields: vec![Vec::from([
                         (
-                            0,
-                            (
-                                IPFixField::SourceIpv4address,
-                                FieldValue::Ip4Addr(Ipv4Addr::new(192, 168, 1, 1)),
-                            ),
+                            IPFixField::IANA(IANAIPFixField::SourceIpv4address),
+                            FieldValue::Ip4Addr(Ipv4Addr::new(192, 168, 1, 1)),
                         ),
                         (
-                            1,
-                            (
-                                IPFixField::DestinationIpv4address,
-                                FieldValue::Ip4Addr(Ipv4Addr::new(192, 168, 1, 2)),
-                            ),
+                            IPFixField::IANA(IANAIPFixField::DestinationIpv4address),
+                            FieldValue::Ip4Addr(Ipv4Addr::new(192, 168, 1, 2)),
                         ),
                         (
-                            2,
-                            (
-                                IPFixField::SourceTransportPort,
-                                FieldValue::DataNumber(DataNumber::U16(1234)),
-                            ),
+                            IPFixField::IANA(IANAIPFixField::SourceTransportPort),
+                            FieldValue::DataNumber(DataNumber::U16(1234)),
                         ),
                         (
-                            3,
-                            (
-                                IPFixField::DestinationTransportPort,
-                                FieldValue::DataNumber(DataNumber::U16(80)),
-                            ),
+                            IPFixField::IANA(IANAIPFixField::DestinationTransportPort),
+                            FieldValue::DataNumber(DataNumber::U16(80)),
                         ),
                         (
-                            4,
-                            (
-                                IPFixField::ProtocolIdentifier,
-                                FieldValue::DataNumber(DataNumber::U8(6)),
-                            ),
+                            IPFixField::IANA(IANAIPFixField::ProtocolIdentifier),
+                            FieldValue::DataNumber(DataNumber::U8(6)),
                         ),
                         (
-                            5,
-                            (
-                                IPFixField::FlowStartSysUpTime,
-                                FieldValue::DataNumber(DataNumber::U32(100)),
-                            ),
+                            IPFixField::IANA(IANAIPFixField::FlowStartSysUpTime),
+                            FieldValue::DataNumber(DataNumber::U32(100)),
                         ),
                         (
-                            6,
-                            (
-                                IPFixField::FlowEndSysUpTime,
-                                FieldValue::DataNumber(DataNumber::U32(200)),
-                            ),
+                            IPFixField::IANA(IANAIPFixField::FlowEndSysUpTime),
+                            FieldValue::DataNumber(DataNumber::U32(200)),
                         ),
                         (
-                            7,
-                            (
-                                IPFixField::SourceMacaddress,
-                                FieldValue::MacAddr("00:00:00:00:00:01".to_string()),
-                            ),
+                            IPFixField::IANA(IANAIPFixField::SourceMacaddress),
+                            FieldValue::MacAddr("00:00:00:00:00:01".to_string()),
                         ),
                         (
-                            8,
-                            (
-                                IPFixField::DestinationMacaddress,
-                                FieldValue::MacAddr("00:00:00:00:00:02".to_string()),
-                            ),
+                            IPFixField::IANA(IANAIPFixField::DestinationMacaddress),
+                            FieldValue::MacAddr("00:00:00:00:00:02".to_string()),
                         ),
                     ])],
                 }),